Improve check placement, test strictness, and pipeline deduplication
- normalize.ml: Promote the full orig_ch.body list (not just List.hd)
  when checks land beyond max_body_level, so entailment has full context.
  Force target >= action_level for checks using primed expressions or
  action params, preventing placement before the action declaration.
- test_check.ml: Replace check_error with explicit collect+check flow
  for CheckWithoutBody test so collection errors fail loudly instead
  of passing silently.
- pipeline.ts: Extract shared buildPantDocument from index.ts and
  helpers.ts, eliminating duplicate pipeline orchestration and
  redundant createSourceFile calls (was 3x, now 1x per invocation).
